Bluefield: Added support for verification of HW rules
Added a new family of classes to the Bluefield test lib to represent
offloaded HW rules, with each type of rule extending the base HwRule class,
and overridding the rules identifying regex.

Added an initial class for L2 forwarding rules.

Added a method to the Bluefield test lib which dumps the HW rules of a
given host for a given gvmi, and verifies that the dumped rules contain
matches for each of a given set of HwRules.

Added a prep script to the repository, which will fetch the pretty_dump
repository used for the HW rule verification, and added the pretty_dump
repository into the gitignore.

Signed-off-by: Gavi Teitz <gavi@mellanox.com>